fix: refactors global initial state + removes spiders
