+++ src/app/pages/global/global-actions.js
@@ -1,9 +1,6 @@
 import { createAction } from 'redux-actions'
-import { createThunkAction } from 'app/utils/redux'
-const { fetch } = window
 
 export const setGlobalSection = createAction('setGlobalSection')
-export const resetGlobalLayers = createAction('resetGlobalLayers')
 export const setWhereToProtectData = createAction('setWhereToProtectData')
 
 export const selectGlobalSelector = createAction(
@@ -19,23 +16,4 @@ export const toggleGlobalLayer = createAction(
   ({ name }) => ({
     analytics: ['global', 'Change Map Layer', name]
   })
-)
-
-export const getWhereToProtectSpiderData = createThunkAction(
-  'getWhereToProtectSpiderData',
-  () => (dispatch, state) => {
-    const classifyScenarios = data =>
-      data.reduce((acc, next) => {
-        const scenario = acc[next.scenario] || []
-        const scenarios = [...scenario, next]
-
-        return { ...acc, [next.scenario]: scenarios }
-      }, {})
-
-    const { whereToProtect: { url } } = state().global
-    return fetch(url)
-      .then(res => res.json())
-      .then(classifyScenarios)
-      .then(data => dispatch(setWhereToProtectData(data)))
-  }
 )
