+++ src/app/pages/global/global-component.jsx
@@ -1,15 +1,10 @@
 import React, { Component } from 'react'
-import SpiderChart from 'components/spider-chart'
 import cx from 'classnames'
-import { Link } from 'react-router-dom'
 import Scroller, { Element as P } from 'components/scroller'
 import NavFooter from 'components/nav-footer'
 import EarthoMeterKnob from 'components/earthometer-knob'
 import Barchart from 'components/barchart'
-
-import styles from './global-styles.scss'
 import uiStyles from 'app/styles/ui'
-import Shapes from './shapes'
 
 class Global extends Component {
   render () {
@@ -18,7 +13,6 @@ class Global extends Component {
       renderToggle,
       renderDropdown,
       className,
-      protectedAnimalsSpider,
       setSection,
       setGlobalSection,
       selectGlobalSelector
@@ -34,91 +28,78 @@ class Global extends Component {
     return (
       <div className={className}>
         <Scroller>
-          <Barchart data={[]} dimension={{ key: 0 }} />
+          <h2>Biodiversity at the planetary scale.</h2>
           <P
             className={uiStyles.slides}
             onScrollFocus={() => updateSections('global:1')}
           >
-            The current extinction rate is many times higher than at any time in
-            Earth’s human history. Without dedicated action that is informed by
-            rigorous evidence, we could incur irreversible loss of species and
-            destroy our living heritage. Smartly managing half the earth for
-            these species will preserve biodiversity and stabilize the Earth’s
-            billion-year-old environmental support system.
+            <span className={uiStyles.innerP}>
+              Species are the fundamental level of study that must drive our
+              understanding of conservation priorities. Regions that are
+              particularly rich in species, or harbor rare species restricted to
+              few locations, are priority areas for conservation.
+            </span>
+            <span>
+              Globally, we have amassed a general, coarse-resolution
+              understanding of species {t('richness')} and
+              {t('rarity')} for {d('global:1', 'birds')}. This information
+              allows us to begin to see the areas most important to manage to
+              protect the bulk of biodiversity. Watch how the map changes as you
+              select different species groups.
+            </span>
+            <Barchart data={[]} dimension={{ key: 0 }} />
           </P>
+
           <P
             className={uiStyles.slides}
             onScrollFocus={() => updateSections('global:2')}
           >
-            Globally, we are able to map species diversity, at least in broad
-            brushes, for an increasing number of {d('global:2', 'birds')}. This
-            allows us to pinpoint important candidate regions for conservation.
-            There are countless more species that haven’t been discovered or
-            formally identified, but they are an integral part of the network of
-            life and are vulnerable to the same threats as the species we have
-            records for.
-            <div className={uiStyles.chartContainer}>
-              <SpiderChart {...protectedAnimalsSpider} />
-              <div className={uiStyles.spiderLegendContainer}>
-                <span className={cx(uiStyles.legend, uiStyles.legendBlue)}>
-                  Percent of species adequately protected
-                </span>
-              </div>
-            </div>
+            This global {t('Protected Areas')} network plays a key role in the
+            conservation of nature and safeguarding of species. However, by
+            overlaying global species {t('richness')} and
+            {t('rarity')} with the {t('Protected Areas')} network, we can see
+            that many species remain insufficiently protected.
+            <Barchart data={[]} dimension={{ key: 0 }} />
           </P>
+
           <P
             className={cx(uiStyles.slides, uiStyles.paragraphAfterChart)}
             onScrollFocus={() => updateSections('global:3')}
           >
-            {t('Protected Areas')} help to maintain the natural balance in the
-            world, and ensure the flow of the many services that ecosystems
-            provide to people, including water, food and raw materials. They
-            play a key role in the conservation of nature and have helped to
-            reduce the current rate of species extinctions. Thanks to global
-            conservation efforts , protected areas have been established in
-            about 15% of the terrestrial landmass. While reserve coverage has
-            been gradually increasing, this trend still falls short from the
-            level required to sustain the world's diversity of life.
-            <Shapes />
+            How can we reduce these conservation gaps and include more species
+            in an expanded network of conservation areas while accounting for
+            the growing constraints from [human press ures], such as road
+            building and urban development?
           </P>
+
           <P
             className={uiStyles.slides}
             onScrollFocus={() => updateSections('global:4')}
           >
+            <span className={uiStyles.innerP}>
+              The Half-Earth Project is integrating multiple layers of high
+              resolution data to identify places where species are
+              insufficiently protected and using this information to guide
+              prioritization of places for conservation.
+            </span>
+            <Barchart data={[]} dimension={{ key: 0 }} />
+            <span className={uiStyles.innerP}>
+              By prioritizing conservation efforts in regions that are rich in
+              species, or regions that have rare species, we can rapidly
+              increase the number of species that have at least minimum
+              conservation protection.
+            </span>
             <EarthoMeterKnob />
-            With basic species maps in hand that can populate a simplified
-            global grid, we can begin to identify conservation gaps and pinpoint
-            the most critical regions to close them. The current reserve system
-            appears to meet minimum conservation targets for about half of the
-            species for the groups analyzed. But this varies among groups, and
-            especially amphibians and cacti are poorly covered. Here, we mimic
-            putative conservation progress by selecting grid cells globally in a
-            sequence that is optimized for meeting conservation targets, and
-            assuming that up to three quarters of a cell could see
-            conservation-focused management.
-            <div className={uiStyles.spiderLegendContainer}>
-              <span className={cx(uiStyles.legend, uiStyles.legendPurple)}>
-                Percent of species meeting conservation target
-              </span>
-            </div>
+            <span className={uiStyles.innerP}>
+              Conservation activities that balance the needs of both humans and
+              nature will require more information, in much finer detail, than
+              what is seen on this prototype Half-Earth map. The goal of the
+              Half-Earth Project is to develop this information, at spatial
+              resolutions of 1km and inclusive of a growing number species, to
+              support biodiversity discovery and conservation worldwide.
+            </span>
           </P>
-          <P
-            className={uiStyles.slides}
-            onScrollFocus={() => updateSections('global:5')}
-          >
-            This simple exercise illustrates the potential for rapid
-            conservation gains through spatial conservation prioritization. It
-            also highlights the importance of including more biodiversity in
-            such planning and illustrates the obvious need for such planning to
-            ultimately happen at a kilometer resolution or finer — as
-            illustrated for{' '}
-            <Link className={styles.link} to="/regional">
-              Southern Africa.
-            </Link>{' '}
-            Half-Earth sets out to help provide the science, tools, and
-            leadership to achieve this vision of a world that consciously
-            safeguards its biodiversity heritage.
-          </P>
+
           <NavFooter from="/regional" to="/" />
         </Scroller>
       </div>
