+++ src/app/pages/global/global-initial-state/global-initial-state.js
@@ -1,0 +1,9 @@
+import layers from './global-layers-initial-state'
+import sections from './global-sections-initial-state'
+import legend from './global-legends-initial-state'
+
+export default {
+  layers,
+  sections,
+  legend
+}
