+++ src/app/pages/global/global.js
@@ -10,22 +10,15 @@ import { actions as cartoActions } from 'providers/carto'
 import { actions as sectionActions } from 'providers/section'
 import * as actions from './global-actions'
 import reducers from './global-reducers'
-import initialState from './global-initial-state'
+import initialState from './global-initial-state/global-initial-state'
 
 class GlobalContainer extends Component {
   constructor (props) {
     super(props)
-    const {
-      getCartoTiles,
-      setGlobalSection,
-      setSection,
-      layers,
-      getWhereToProtectSpiderData
-    } = props
+    const { getCartoTiles, setGlobalSection, setSection, layers } = props
     requestCartos({ layers, getCartoTiles })
     setGlobalSection('global:1')
     setSection('global:1')
-    getWhereToProtectSpiderData()
   }
   render () {
     return createElement(GlobalComponent, assign(this.props))
@@ -36,45 +29,10 @@ const mapStateToProps = ({
   map,
   global,
   section,
-  earthSaved,
   getWhereToProtectSpiderData
 }) => {
-  const index = Math.round(earthSaved.value)
-  const whereToProtectSpider = {
-    dimensions: global.whereToProtect.dimensions,
-    data: (global.whereToProtect.data[index] || [])
-      .filter(d => d.taxa !== 'all')
-      .map(({ taxa, ...rest }) => ({
-        ...rest,
-        subject: taxa.toUpperCase(),
-        tooltip: [
-          {
-            label: '%',
-            value: rest.percentSpeciesMeetingTargetProtectedAreaViaAny,
-            color: '#8366e4'
-          }
-        ]
-      }))
-  }
-
-  const protectedAnimalsSpider = {
-    ...global.protectedAnimals,
-    data: global.protectedAnimals.data.map(d => ({
-      ...d,
-      tooltip: [
-        {
-          value: d.percent,
-          label: '%',
-          color: '#3850d6'
-        }
-      ]
-    }))
-  }
-
   return {
     map,
-    whereToProtectSpider,
-    protectedAnimalsSpider,
     getWhereToProtectSpiderData,
     layers: global.layers,
     section: section.section,
